+++ QuanLyThuVien/Models/Librarian.cs
@@ -18,6 +18,7 @@ namespace QuanLyThuVien.Models {
 
         [Required(ErrorMessage ="Please confirm password.")]
         [DataType(DataType.Password)]
+        [Compare("Password", ErrorMessage = "The password and confirmation password do not match.")]
         public string ConfirmPassword { get; set; }
 
         [Required(ErrorMessage ="First Name is required.")]