+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+[0.1.1] 19/04/2016
+
+## Added
+- Add module Delete Librarian
+
+## Changed
+- Change Session logic in LibraryController
+
 [0.1] 16/04/2016
 
 ## Added